Broke large nnscore_helper file into nnscore_pdb and nnscore_utils.
